+++ storage/index.js
@@ -1,0 +1,9 @@
+/**
+ * storage surface exports adapters and working-set helpers without selecting a vendor backend.
+ */
+export * from "./adapter.js";
+export * from "./cache.js";
+export * from "./checksum.js";
+export * from "./chunked.js";
+export * from "./content.js";
+export * from "./sync.js";